# docs/swagger.yaml
definitions:
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.SuccessResponse:
    properties:
      data:
        description: Опциональные данные
      message:
        type: string
    type: object
...
  models.Grade:
    properties:
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      student_id:
        type: integer
      subject_id:
        type: integer
      updated_at:
        type: string
      value:
        description: Оценка (2-5)
        type: integer
    type: object
  models.GradeDetail:
    properties:
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      student_id:
        type: integer
      student_name:
        type: string
      subject_name:
        type: string
      updated_at:
        type: string
      value:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/grades:
    post:
      consumes:
      - application/json
      description: Создаёт новую запись об оценке
      parameters:
      - description: Данные об оценке
        in: body
        name: grade
        required: true
        schema:
          $ref: '#/definitions/models.Grade'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Создать оценку
      tags:
      - Grades
  /api/grades/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет запись об оценке
      parameters:
      - description: ID оценки
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Удалить оценку
      tags:
      - Grades
    put:
      consumes:
      - application/json
      description: Обновляет существующую запись об оценке
      parameters:
      - description: ID оценки
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные об оценке
        in: body
        name: grade
        required: true
        schema:
          $ref: '#/definitions/models.Grade'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Обновить оценку
      tags:
      - Grades
  /api/grades/group/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает список оценок для конкретной группы
      parameters:
      - description: ID группы
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            items:
              $ref: '#/definitions/models.GradeDetail'
            type: array
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Оценки не найдены
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить оценки группы по её ID
      tags:
      - Grades
  /api/grades/student/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает список оценок для конкретного студента
      parameters:
      - description: ID студента
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            items:
              $ref: '#/definitions/models.GradeDetail'
            type: array
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Оценки не найдены
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить оценки студента по его ID
      tags:
      - Grades
...
swagger: "2.0"